Заготовка под главу 7
