+++ .gitignore
@@ -5,5 +5,4 @@ venv/
 docs/
 
 # .env-файлы
-/code/01_quickstart/.env
-/code/06_special_updates/.env
+.env
