--- code/04_fsm/README.md
@@ -1,1 +1,0 @@
-Примечание: не забудьте сделать копию файла `bot.ini.example` под именем `bot.ini` и далее работать с этой копией!
