--- code/04_fsm/app/handlers/common.py
@@ -1,29 +1,0 @@
-from aiogram import Dispatcher, types
-from aiogram.dispatcher import FSMContext
-from aiogram.dispatcher.filters import Text, IDFilter
-
-
-async def cmd_start(message: types.Message, state: FSMContext):
-    await state.finish()
-    await message.answer(
-        "Выберите, что хотите заказать: напитки (/drinks) или блюда (/food).",
-        reply_markup=types.ReplyKeyboardRemove()
-    )
-
-
-async def cmd_cancel(message: types.Message, state: FSMContext):
-    await state.finish()
-    await message.answer("Действие отменено", reply_markup=types.ReplyKeyboardRemove())
-
-
-# Просто функция, которая доступна только администратору,
-# чей ID указан в файле конфигурации.
-async def secret_command(message: types.Message):
-    await message.answer("Поздравляю! Эта команда доступна только администратору бота.")
-
-
-def register_handlers_common(dp: Dispatcher, admin_id: int):
-    dp.register_message_handler(cmd_start, commands="start", state="*")
-    dp.register_message_handler(cmd_cancel, commands="cancel", state="*")
-    dp.register_message_handler(cmd_cancel, Text(equals="отмена", ignore_case=True), state="*")
-    dp.register_message_handler(secret_command, IDFilter(user_id=admin_id), commands="abracadabra")
