--- code/04_fsm/app/handlers/food.py
@@ -1,51 +1,0 @@
-from aiogram import Dispatcher, types
-from aiogram.dispatcher import FSMContext
-from aiogram.dispatcher.filters.state import State, StatesGroup
-
-
-available_food_names = ["суши", "спагетти", "хачапури"]
-available_food_sizes = ["маленькую", "среднюю", "большую"]
-
-
-class OrderFood(StatesGroup):
-    waiting_for_food_name = State()
-    waiting_for_food_size = State()
-
-
-async def food_start(message: types.Message):
-    keyboard = types.ReplyKeyboardMarkup(resize_keyboard=True)
-    for name in available_food_names:
-        keyboard.add(name)
-    await message.answer("Выберите блюдо:", reply_markup=keyboard)
-    await OrderFood.waiting_for_food_name.set()
-
-
-# Обратите внимание: есть второй аргумент
-async def food_chosen(message: types.Message, state: FSMContext):
-    if message.text.lower() not in available_food_names:
-        await message.answer("Пожалуйста, выберите блюдо, используя клавиатуру ниже.")
-        return
-    await state.update_data(chosen_food=message.text.lower())
-
-    keyboard = types.ReplyKeyboardMarkup(resize_keyboard=True)
-    for size in available_food_sizes:
-        keyboard.add(size)
-    # Для простых шагов можно не указывать название состояния, обходясь next()
-    await OrderFood.next()
-    await message.answer("Теперь выберите размер порции:", reply_markup=keyboard)
-
-
-async def food_size_chosen(message: types.Message, state: FSMContext):
-    if message.text.lower() not in available_food_sizes:
-        await message.answer("Пожалуйста, выберите размер порции, используя клавиатуру ниже.")
-        return
-    user_data = await state.get_data()
-    await message.answer(f"Вы заказали {message.text.lower()} порцию {user_data['chosen_food']}.\n"
-                         f"Попробуйте теперь заказать напитки: /drinks", reply_markup=types.ReplyKeyboardRemove())
-    await state.finish()
-
-
-def register_handlers_food(dp: Dispatcher):
-    dp.register_message_handler(food_start, commands="food", state="*")
-    dp.register_message_handler(food_chosen, state=OrderFood.waiting_for_food_name)
-    dp.register_message_handler(food_size_chosen, state=OrderFood.waiting_for_food_size)
