--- code/04_fsm/config/bot.ini.example
@@ -1,5 +1,0 @@
-[tg_bot]
-token = 1234567890:abcdefghijklmnopqrstuvwxyz
-admin_id = 987654321
-
-# Не забудьте скопировать этот файл под именем bot.ini
