--- code/04_fsm/requirements.txt
@@ -1,1 +1,0 @@
-aiogram==2.11.2
