+++ code/07_fsm/bot.py
@@ -1,0 +1,22 @@
+import asyncio
+import logging
+
+from aiogram import Bot, Dispatcher
+
+from config_reader import config
+
+
+async def main():
+    logging.basicConfig(
+        level=logging.INFO,
+        format="%(asctime)s - %(levelname)s - %(name)s - %(message)s",
+    )
+
+    dp = Dispatcher()
+    bot = Bot(config.bot_token.get_secret_value(), parse_mode="HTML")
+
+    await dp.start_polling(bot, allowed_updates=dp.resolve_used_update_types())
+
+
+if __name__ == '__main__':
+    asyncio.run(main())
