+++ code/07_fsm/config_reader.py
@@ -1,0 +1,12 @@
+from pydantic import BaseSettings, SecretStr
+
+
+class Settings(BaseSettings):
+    bot_token: SecretStr
+
+    class Config:
+        env_file = '.env'
+        env_file_encoding = 'utf-8'
+
+
+config = Settings()
