+++ code/07_fsm/env_dist
@@ -1,0 +1,3 @@
+# Не забудьте скопировать этот файл под именем .env
+
+BOT_TOKEN = 0000000000:AaBbCcDdEeFfGgHhIiJjKkLlMmNn
