+++ code/07_fsm/requirements.txt
@@ -1,0 +1,16 @@
+aiofiles==0.8.0
+aiogram==3.0.0b3
+aiohttp==3.8.1
+aiosignal==1.2.0
+async-timeout==4.0.2
+attrs==22.1.0
+charset-normalizer==2.1.0
+frozenlist==1.3.1
+idna==3.3
+magic-filter==1.0.8
+multidict==6.0.2
+pydantic==1.9.1
+Pygments==2.12.0
+python-dotenv==0.20.0
+typing_extensions==4.3.0
+yarl==1.8.1
